@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '@tms/contexts/AuthContext';
 import './Auth.css';
 
 const Login = ({ onSwitchToRegister }) => {
