@@ -8,58 +8,77 @@ const getSlackMessage = ({
     skipped,
     duration,
     result,
-}) => {1
-    //slack block kit
+}) => {
+    // 결과에 따른 색상 결정
+    const color = failed > 0 ? '#ff0000' : '#36a64f'; // 실패 시 빨간색, 성공 시 초록색
+    const statusEmoji = failed > 0 ? '❌' : '✅';
+    
+    // 슬랙 블록 킷 구성
+    const blocks: any[] = [
+        {
+            type: 'header',
+            text: {
+                type: 'plain_text',
+                text: `${statusEmoji} Playwright Test Report`,
+            },
+        },
+        {
+            type: 'section',
+            fields: [
+                {
+                    type: 'mrkdwn',
+                    text: `*총 테스트 수:*\n${all}개`,
+                },
+                {
+                    type: 'mrkdwn',
+                    text: `*실행 시간:*\n${duration}`,
+                },
+            ],
+        },
+        {
+            type: 'section',
+            fields: [
+                {
+                    type: 'mrkdwn',
+                    text: `*✅ 통과:*\n${passed}개`,
+                },
+                {
+                    type: 'mrkdwn',
+                    text: `*❌ 실패:*\n${failed}개`,
+                },
+                {
+                    type: 'mrkdwn',
+                    text: `*⚠️ 스킵:*\n${skipped}개`,
+                },
+            ],
+        },
+    ];
+    
+    // 실패한 테스트가 있는 경우 상세 정보 추가
+    if (failed > 0 && result) {
+        blocks.push({
+            type: 'section',
+            text: {
+                type: 'mrkdwn',
+                text: `*실패한 테스트 상세:*\n\`\`\`${result}\`\`\``,
+            },
+        });
+    } else if (failed === 0) {
+        blocks.push({
+            type: 'section',
+            text: {
+                type: 'mrkdwn',
+                text: '*🎉 모든 테스트가 통과했습니다!*',
+            },
+        });
+    }
+    
     return {
-        blocks: [
+        blocks,
+        attachments: [
             {
-                type: 'header',
-                text: {
-                    type: 'plain_text',
-                    text: 'Playwright Test Report',
-                },
-            },
-            {
-                type: 'section',
-                text: {
-                    type: 'mrkdwn',
-                    text: `*Total Tests: ${all}*`,
-                },
-            },
-            {
-                type: 'section',
-                text: {
-                    type: 'mrkdwn',
-                    text: `*Passed Tests: ${passed}*`,
-                },
-            },
-            {
-                type: 'section',
-                text: {
-                    type: 'mrkdwn',
-                    text: `*Failed Tests: ${failed}*`,
-                },
-            },
-            {
-                type: 'section',
-                text: {
-                    type: 'mrkdwn',
-                    text: `*Skipped Tests: ${skipped}*`,
-                },
-            },
-            {
-                type: 'section',
-                text: {
-                    type: 'mrkdwn',
-                    text: `*Duration: ${duration}*`,
-                },
-            },
-            {
-                type: 'section',
-                text: {
-                    type: 'mrkdwn',
-                    text: `*Result: ${result}*`,
-                },
+                color,
+                blocks: [],
             },
         ],
     };
@@ -72,12 +91,12 @@ class MyReporter implements Reporter {
     skipped = 0;
     failMessages = '';
 
-    onBeing(_: FullConfig, suite: Suite) {
+    onBegin(_: FullConfig, suite: Suite) {
         this.all = suite.allTests().length;
     }
 
     onTestEnd(test: TestCase, result: TestResult) {
-        const testDuration = '${result.duration / 1000).toFixed(1)}s';
+        const testDuration = `${(result.duration / 1000).toFixed(1)}s`;
         const fileName = path.basename(test.location.file);
         const testTitle = test.title;
 
@@ -105,7 +124,7 @@ class MyReporter implements Reporter {
 
     async onEnd(result: FullResult) {
         const blockKit = await this.getBlockKit(result);
-        const webhookUrl = await process.env.SLACK_WEBHOOK_URL;
+        const webhookUrl = process.env.SLACK_WEBHOOK_URL;
 
         if (!webhookUrl) {
             console.error('SLACK_WEBHOOK_URL 환경 변수가 설정되지 않았습니다.');
